// src/app/figures/types/all-calculation-data.ts
export interface AllCalculationData {
    sideALength?: number;
    sideBLength?: number;
    sideCLength?: number;
    calculationType?: string;
    shape?: string;
    checkTriangle?: boolean;
    nameALength?: string;
    nameBLength?: string;
    nameCLength?: string;
    resNameALength?: string;
    resNameBLength?: string;
    resNameCLength?: string;
    shapeName?: string;
    calculationName?: string;
  }